The line flags on IRIX are wider than they are on Linux or FreeBSD. Fixed coding standard violations while I am at it.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@13779 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vapor/vpr/md/POSIX/IO/Port/SerialPortImplTermios.h
@@ -1116,10 +1116,10 @@ protected:
                             const bool print_sys_err = true);
 
    /** Helper for setting line status flags */
-   vpr::ReturnStatus setLineFlag(Uint8 flag, bool val);
+   vpr::ReturnStatus setLineFlag(Uint16 flag, bool val);
 
    /** Helper for getting lines status flags */
-   bool getLineFlag(Uint8 flag) const;
+   bool getLineFlag(Uint16 flag) const;
 
    /**
     * Converts a termios baud rate to its corresponding integer value.
